Aterações provisórias para revisão
